Remove unnecessary projects. Move IntroNativeMemory to Samples.
