--- tests/BenchmarkDotNet.IntegrationTests.Native/stdafx.cpp
@@ -1,1 +1,0 @@
-#include "stdafx.h"
