--- tests/BenchmarkDotNet.IntegrationTests.NativeWraper/BenchmarkDotNet.IntegrationTests.NativeWraper.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\..\build\common.props" />
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Condition="Exists('$(MSBuildThisFileDirectory)..\..\x64\Release\BenchmarkDotNet.IntegrationTests.Native.dll')" Include="$(MSBuildThisFileDirectory)..\..\x64\Release\BenchmarkDotNet.IntegrationTests.Native.dll">
-      <Link>BenchmarkDotNet.IntegrationTests.Native.dll</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <Visible>False</Visible>
-    </None>
-   </ItemGroup>
-</Project>
